moved htlc related code into htlc.go
